refactor write_file.py

+++ functions/write_file.py
@@ -1,13 +1,20 @@
 import os
 def write_file(working_directory, file_path, content):
-    try:
-        full_file_path = os.path.join(working_directory,file_path)
-        abs_file_path = os.path.abspath(full_file_path)
-        abs_dir_path = os.path.abspath(working_directory)
+    full_file_path = os.path.join(working_directory,file_path)
+    abs_file_path = os.path.abspath(full_file_path)
+    abs_dir_path = os.path.abspath(working_directory)
 
-        if not abs_file_path.startswith(abs_dir_path):
-            return f'Error: Cannot write to "{file_path}" as it is outside the permitted working directory'
-        os.makedirs(os.path.dirname(abs_file_path), exist_ok = True)
+    if not abs_file_path.startswith(abs_dir_path):
+        return f'Error: Cannot write to "{file_path}" as it is outside the permitted working directory'
+
+    if not os.path.exists(abs_file_path):
+        try:
+            os.makedirs(os.path.dirname(abs_file_path), exist_ok = True)
+        except Exception as e:
+            return f"Error: creating directory: {e}"
+    if os.path.exists(abs_file_path) and os.path.isdir(abs_file_path):
+        return f'Error: "{file_path}" is a directory, not a file'
+    try:
         with open(abs_file_path, "w") as f:
             f.write(content)
             return f'Successfully wrote to "{file_path}" ({len(content)} characters written)'
